none of this tested yet due to time constraints.
  Added code to check for a bit already set in allocator bitmap.
The old code would clear a bit that possibly could have been 0 already,
and then decrement in use counter, thus prematurely freeing a block
that was actually still in use. Freeing the same block twice should not
happen, but if it =does= there is no point trashing the block allocator
because of it. This can happen if exit_client() is called twice on same
cptr for example. Added debug and SYSLOG option for
SYSLOG_BLOCK_ALLOCATOR. Files affected, list.c, blalloc.c, config.h.dist

  Taner had removed debug warning notice and DEBUG warning
of a cptr not found in hash table. This means that an exit_client()
called twice on same cptr would not be flagged as early as it could have
been. I restored the original code. Files affected s_misc.c

  I replaced a call to MyRealloc() simply because I do not trust all
OS's version of realloc(). It has undefined results depending on which
OS you are using in the case of a 0 being used for the new size.
I just replaced with a free/malloc combination. This is in the m_away
code. Files affected s_user.c

  Fixed the new kline processing code, in some odd cases it looks like
a cptr could have been asked to exit_client() twice. This would have
caused the block allocator to be freaked out, and opers to start
getting odd notices from list.c
This one:

      sendto_ops("* %#x user (%s!%s@%s) %#x %#x %#x %d %d *",
		 cptr, cptr ? cptr->name : "<noname>",
		 user->username, user->host, user,
		 user->invited, user->channel, user->joined,
		 user->refcnt);

Files affected ircd.c